+++ core/server/api_container/server/startosis_engine/startosis_runner.go
@@ -2,6 +2,7 @@ package startosis_engine
 
 import (
 	"context"
+	"fmt"
 	"github.com/kurtosis-tech/kurtosis/api/golang/core/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/binding_constructors"
 	"github.com/sirupsen/logrus"
@@ -21,6 +22,8 @@ const (
 	startingInterpretationMsg = "Interpreting Starlark code - execution will begin shortly"
 	startingValidationMsg     = "Pre-validating Starlark code and downloading docker images - execution will begin shortly"
 	startingExecutionMsg      = "Starting execution"
+
+	missingRunMethodErrorFromStarlark = "Evaluation error: module has no .run field or method\n\tat [3:12]: <toplevel>"
 )
 
 func NewStartosisRunner(interpreter *StartosisInterpreter, validator *StartosisValidator, executor *StartosisExecutor) *StartosisRunner {
@@ -45,6 +48,7 @@ func (runner *StartosisRunner) Run(ctx context.Context, dryRun bool, moduleId st
 
 		instructionsList, interpretationError := runner.startosisInterpreter.Interpret(ctx, moduleId, serializedStartosis, serializedParams)
 		if interpretationError != nil {
+			interpretationError = maybeMakeMissingRunMethodErrorFriendlier(interpretationError, moduleId)
 			kurtosisExecutionResponseLines <- binding_constructors.NewKurtosisExecutionResponseLineFromInterpretationError(interpretationError)
 			return
 		}
@@ -58,7 +62,10 @@ func (runner *StartosisRunner) Run(ctx context.Context, dryRun bool, moduleId st
 		kurtosisExecutionResponseLines <- progressInfo
 
 		validationErrorsChan := runner.startosisValidator.Validate(ctx, instructionsList)
-		forwardKurtosisResponseLineChannelUntilSourceIsClosed(validationErrorsChan, kurtosisExecutionResponseLines)
+		if messagesWereReceived := forwardKurtosisResponseLineChannelUntilSourceIsClosed(validationErrorsChan, kurtosisExecutionResponseLines); messagesWereReceived {
+			return
+		}
+
 		logrus.Debugf("Successfully validated Startosis script")
 
 		// Execution starts > send progress info. This will soon be overridden byt the first instruction execution
@@ -73,10 +80,20 @@ func (runner *StartosisRunner) Run(ctx context.Context, dryRun bool, moduleId st
 	return kurtosisExecutionResponseLines
 }
 
-func forwardKurtosisResponseLineChannelUntilSourceIsClosed(sourceChan <-chan *kurtosis_core_rpc_api_bindings.KurtosisExecutionResponseLine, destChan chan<- *kurtosis_core_rpc_api_bindings.KurtosisExecutionResponseLine) {
+func forwardKurtosisResponseLineChannelUntilSourceIsClosed(sourceChan <-chan *kurtosis_core_rpc_api_bindings.KurtosisExecutionResponseLine, destChan chan<- *kurtosis_core_rpc_api_bindings.KurtosisExecutionResponseLine) bool {
+	messagesWereReceived := false
 	for executionResponseLine := range sourceChan {
 		logrus.Debugf("Received kurtosis execution line Kurtosis:\n%v", executionResponseLine)
 		destChan <- executionResponseLine
+		messagesWereReceived = true
 	}
 	logrus.Debug("Kurtosis instructions stream was closed. Exiting execution loop")
+	return messagesWereReceived
+}
+
+func maybeMakeMissingRunMethodErrorFriendlier(originalError *kurtosis_core_rpc_api_bindings.KurtosisInterpretationError, moduleId string) *kurtosis_core_rpc_api_bindings.KurtosisInterpretationError {
+	if originalError.GetErrorMessage() == missingRunMethodErrorFromStarlark {
+		return binding_constructors.NewKurtosisInterpretationError(fmt.Sprintf("No 'run' function found in file '%v/main.star'; a 'run' entrypoint function is required in the main.star file of any Kurtosis package", moduleId))
+	}
+	return originalError
 }